Add weather tests back.
